requirements.txt
flask==1.0.0
requests==2.18.4
pycryptodome==3.10.1
